requirements-min.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements-min.txt requirements-min.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
cryptography==46.0.3
    # via
    #   pdfminer-six
    #   pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   -r requirements-min.in
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.7.0
    # via yarl
packaging==25.0
    # via
    #   deprecation
    #   pytesseract
pdfminer-six==20250506
    # via -r requirements-min.in
pillow==11.3.0
    # via
    #   -r requirements-min.in
    #   pytesseract
    #   ttkbootstrap
postgrest==2.22.0
    # via supabase
propcache==0.4.1
    # via yarl
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   postgrest
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.4
    # via pydantic
pyjwt[crypto]==2.10.1
    # via supabase-auth
pymupdf==1.26.5
    # via -r requirements-min.in
pypdf==6.1.1
    # via -r requirements-min.in
pytesseract==0.3.13
    # via -r requirements-min.in
python-dotenv==1.1.1
    # via -r requirements-min.in
pyyaml==6.0.3
    # via -r requirements-min.in
realtime==2.22.0
    # via supabase
requests==2.32.5
    # via -r requirements-min.in
sniffio==1.3.1
    # via anyio
storage3==2.22.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
supabase==2.22.0
    # via -r requirements-min.in
supabase-auth==2.22.0
    # via supabase
supabase-functions==2.22.0
    # via supabase
ttkbootstrap==1.14.7
    # via -r requirements-min.in
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via
    #   -r requirements-min.in
    #   requests
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase-functions
